test: remove low-value test files
Remove tests that are testing impossible scenarios, trivial functions,
or duplicating validation already covered at the client layer:

- configure_wrong_type_test.go: Tests ProviderData type errors
  (Terraform controls this, can't happen in real usage)
- provider_base_url_test.go: Tests a 5-line env var fallback function
- dns_records_validators_test.go: Duplicates client-layer validation tests
- records/a_record_validator_test.go & srv_record_validator_test.go:
  Test that thin wrapper calls through (already covered at client layer)
- dns_records_resource_crud_test.go: Reimplements acceptance tests
- Error-path tests (dns_records_resource_errors_test.go, etc.):
  Use fragile call-counter logic that breaks with implementation changes
- domain_resource_helpers_test.go: Overlaps with existing tests
- provider_configure_test.go: Remove resolveString tests that duplicate
  domain_common_test.go tests

Keep data_sources tests (data_sources_unit_test.go, data_sources_errors_test.go)
and domain_resource_import_test.go as they test real provider functionality.

This reduces test code from ~3045 to ~1100 lines while keeping meaningful
coverage and error path tests that add value.
